restructure: merged feed files
